+++ doc/whats_new.rst
@@ -87,6 +87,9 @@ API changes summary
      models. This allows to have a regularization parameter independent
      of the number of samples. The scale_C parameter will disappear in v0.12.
 
+   - In :class:`svm.LinearSVC`, the `multi_class` parameter was renamed
+     to `crammer_singer` to avoid confusion.
+
 .. _changes_0_10:
 
 0.10